requirements.txt
torch>=2.0.0
transformers>=4.34.0
sentence-transformers
faiss-cpu
PyMuPDF
openai>=1.0.0
streamlit
numpy
